[treewide] lora: use int16_t for RSSI value [backport 2022.01]

+++ drivers/sx126x/sx126x_netdev.c
@@ -291,9 +291,9 @@ static int _get(netdev_t *netdev, netopt_t opt, void *val, size_t max_len)
         return sizeof(netopt_enable_t);
 
     case NETOPT_RSSI:
-        assert(max_len >= sizeof(int8_t));
+        assert(max_len >= sizeof(int16_t));
         sx126x_get_rssi_inst(dev, ((int16_t *)val));
-        return sizeof(int8_t);
+        return sizeof(int16_t);
 
     default:
         break;
