+++ drivers/sx127x/sx127x_netdev.c
@@ -345,9 +345,9 @@ static int _get(netdev_t *netdev, netopt_t opt, void *val, size_t max_len)
         return sizeof(netopt_enable_t);
 
     case NETOPT_RSSI:
-        assert(max_len >= sizeof(int8_t));
-        *((int8_t *)val) = sx127x_read_rssi(dev);
-        return sizeof(int8_t);
+        assert(max_len >= sizeof(int16_t));
+        *((int16_t *)val) = sx127x_read_rssi(dev);
+        return sizeof(int16_t);
 
     default:
         break;
